Move watchdog logic to new separate file
